create separate component for NewsletterPrivacy message
