+++ dotcom-rendering/src/web/components/SecureSignup.tsx
@@ -1,81 +1,19 @@
 import createCache from '@emotion/cache';
 import { CacheProvider, css } from '@emotion/react';
 import createEmotionServer from '@emotion/server/create-instance';
-import { neutral, space, text, textSans } from '@guardian/source-foundations';
-import {
-	Button,
-	Label,
-	Link,
-	TextInput,
-} from '@guardian/source-react-components';
+import { neutral, space } from '@guardian/source-foundations';
+import { Button, Label, TextInput } from '@guardian/source-react-components';
 import { renderToString } from 'react-dom/server';
 import { Island } from './Island';
+import { NewsletterPrivacyMessage } from './NewsletterPrivacyMessage';
 import { SecureSignupIframe } from './SecureSignupIframe.importable';
-
-// The Google documentation specifies that if the 'recaptcha-badge' is hidden,
-// their T+C's must be displayed instead. <SecureSignupIframe> hides the
-// badge, so <RecaptchaTerms> must be displayed with it.
-// https://developers.google.com/recaptcha/docs/faq#id-like-to-hide-the-recaptcha-badge.-what-is-allowed
 
 type Props = { newsletterId: string; successDescription: string };
 
-const termsStyle = css`
-	${textSans.xxsmall()}
-	color: ${text.supporting};
-	a {
-		${textSans.xxsmall({ fontWeight: 'bold' })};
-		color: ${neutral[0]};
-		text-decoration: underline;
-		:hover {
-			color: ${neutral[0]};
-			text-decoration: underline;
-		}
-	}
-	strong {
-		color: ${neutral[0]};
-		font-weight: bold;
-	}
-`;
-
-const PrivacyTerms = () => {
-	return (
-		<span css={termsStyle}>
-			<strong>Privacy Notice: </strong>
-			Newsletters may contain info about charities, online ads, and
-			content funded by outside parties. For more information see our{' '}
-			<Link
-				data-ignore="global-link-styling"
-				href="https://www.theguardian.com/help/privacy-policy"
-				rel="noopener noreferrer"
-			>
-				privacy policy
-			</Link>
-			.&nbsp;
-		</span>
-	);
-};
-
-const RecaptchaTerms = () => (
-	<span css={termsStyle}>
-		We use Google reCaptcha to protect our website and the Google{' '}
-		<Link
-			data-ignore="global-link-styling"
-			href="https://policies.google.com/privacy"
-			rel="noopener noreferrer"
-		>
-			Privacy Policy
-		</Link>{' '}
-		and{' '}
-		<Link
-			data-ignore="global-link-styling"
-			href="https://policies.google.com/terms"
-			rel="noopener noreferrer"
-		>
-			Terms of Service
-		</Link>{' '}
-		apply.
-	</span>
-);
+// The Google documentation specifies that if the 'recaptcha-badge' is hidden,
+// their T+C's must be displayed instead. <SecureSignupIframe> hides the
+// badge, so <NewsletterPrivacyMessage> must be displayed with it.
+// https://developers.google.com/recaptcha/docs/faq#id-like-to-hide-the-recaptcha-badge.-what-is-allowed
 
 /**
  * This function renders the content to be used within the iframe,
@@ -176,8 +114,7 @@ export const SecureSignup = ({ newsletterId, successDescription }: Props) => {
 					margin-top: ${space[2]}px;
 				`}
 			>
-				<PrivacyTerms />
-				<RecaptchaTerms />
+				<NewsletterPrivacyMessage />
 			</div>
 		</>
 	);
